add support for hugeint to TransformPythonValue and always try to use the smallest integer type

+++ tools/pythonpkg/tests/fast/pandas/test_df_object_resolution.py
@@ -19,7 +19,7 @@ class TestResolveObjectColumns(object):
         data = [5, 0, 3]
         df_in = create_generic_dataframe(data)
         # These are float64 because pandas would force these to be float64 even if we set them to int8, int16, int32, int64 respectively
-        df_expected_res = pd.DataFrame({'0': pd.Series(data=data)})
+        df_expected_res = pd.DataFrame({'0': pd.Series(data=data, dtype='int8')})
         df_out = duckdb.query_df(df_in, "data", "SELECT * FROM data").df()
         print(df_out)
         pd.testing.assert_frame_equal(df_expected_res, df_out)
